Рефакторинг авторизации
