+++ loginDir/checker.py
@@ -1,3 +1,6 @@
+import data
+
+
 class Checker:
 	def __init__(self, username, password):
 		self.username = username
@@ -7,15 +10,15 @@ class Checker:
 		self.checkUserNameLogin()
 		self.checkPassword()
 
-
 	def checkUserNameLogin(self):
-		LoginFind = Workers.get(self.username)
+		LoginFind = data.Workers.get(self.username)
 		if LoginFind:
 			self.LoginChecker = True
 
 	def checkPassword(self):
 		if self.password:
-			if 'password' in Workers[self.username] and Workers[self.username]['password'] == self.password:
-					self.PasswordChecker = True
+			if 'password' in data.Workers[self.username] and data.Workers[self.username]['password'] == self.password:
+				self.PasswordChecker = True
+
 	def __str__(self):
 		return 'Пользователь %s' % self.username
